+++ CHANGES.rst
@@ -11,6 +11,10 @@ CHANGES
 
 - Added Python 3.10 and 3.11 support.
 
+- Removed Python 2 support.
+
+- Fixed Deadline / ExceptionName parsing.
+
 
 4.3.0 (2022-04-07)
 ------------------